+++ src/main/java/dev/jpp/listapi/entity/UserRole.java
@@ -15,7 +15,7 @@ public class UserRole {
 	@Id
 	@GeneratedValue
 	@Column(name="id")
-	private int id;
+	private Long id;
 	
 	@ManyToOne(fetch=FetchType.EAGER)
 	@JoinColumn(name="user_id", nullable=false)
@@ -28,18 +28,18 @@ public class UserRole {
 		super();
 	}
 	
-	public UserRole(int id, User user, String role) {
+	public UserRole(Long id, User user, String role) {
 		super();
 		this.id = id;
 		this.user = user;
 		this.role = role;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
